+++ groups/btl/btlsos/btlsos_tcptimedchannel.cpp
@@ -80,7 +80,7 @@ int adjustVecBuffer(const VECTYPE        *buffers,
                 (char*) const_cast<void *>(buffers[i].buffer()),
                 buffers[i].length()));
     }
-    return vector->size();
+    return static_cast<int>(vector->size());
 }
 
 namespace btlsos {
@@ -2372,7 +2372,7 @@ int TcpTimedChannel::writev(int *augStatus, int length, int flags)
 
     int rc = 0;
     int numBytesWritten = 0;
-    int originalNumBuffers = d_ovecBuffers.size();
+    int originalNumBuffers = static_cast<int>(d_ovecBuffers.size());
     int numBuffers = originalNumBuffers;
     bsl::vector<btls::Ovec> *writeBuffers = &d_ovecBuffers;
     
@@ -2434,7 +2434,7 @@ int TcpTimedChannel::timedWritev(int                       *augStatus,
     
     int numBytesWritten = 0;
     int retValue = 0;
-    int originalNumBuffers = d_ovecBuffers.size();
+    int originalNumBuffers = static_cast<int>(d_ovecBuffers.size());
     int numBuffers = originalNumBuffers;
     bsl::vector<btls::Ovec> *writeBuffers = &d_ovecBuffers;
